Implement QT breakdown of image into QT nodes. We are so awesome.
